requirements.txt
langchain-google-genai
langchain-groq
python-dotenv
jupyter
notebook
langchain
requests
fastapi
uvicorn
chromadb
pysqlite3-binary
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6